AgentLearn/MiniCoderPlus/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pywinpty>=2.0.0
websockets>=11.0.0
